% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{all_county_fips}
\alias{all_county_fips}
\title{fips of each county}
\format{A data.table with 3109 rows and 3 variables
\describe{
  \item{fips}{fips codes of state}
  \item{state}{lower case state names}
  \item{county}{lower case county names}

}}
\usage{
all_county_fips
}
\description{
fips of each county
}
\keyword{datasets}
