% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{all_state_fips}
\alias{all_state_fips}
\title{fips of each state}
\format{A data.table with 53 rows and 3 variables
\describe{
  \item{fips}{fips codes of state}
  \item{abbr}{abbreviations of state}
  \item{state}{lower case state names}

}}
\usage{
all_state_fips
}
\description{
fips of each state
}
\keyword{datasets}
