% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{all_zipcode}
\alias{all_zipcode}
\title{all zip code in each county}
\format{A data.table with 172 rows and 2 variables
\describe{
  \item{abbr}{abbreviation of state}
  \item{county}{county name}
  \item{ZCTA5}{ZCTA5}
}}
\usage{
all_zipcode
}
\description{
all zip code in each county
}
\keyword{datasets}
