% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_census.R
\name{geom_census}
\alias{geom_census}
\title{Plot census data on ggmap}
\usage{
geom_census(survey, year, table_contents, summary_level, states = NULL,
  areas = NULL, N = 100, mapping = NULL, data = NULL, geom = "point",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{survey}{"decennial" for decennial census, "acs5year" for ACS 5-year
estimates, "acs1year" for ACS 1-year estimate}

\item{year}{year of the survey}

\item{table_contents}{contents of table to be read}

\item{summary_level}{select which summary level to keep, "*" to keep all. It
takes strings including "state", "county", "county subdivision", "place",
"tract", "block group", and "block" for the most common levels. It also
take code for level.}

\item{states}{abbreviation of a state of which data to be read, such as "RI".}

\item{areas}{areas of which data to be read}

\item{N}{integer when state is NULL, N x N grid points on the view of map are
used to determine the states and counties within the view.}

\item{mapping}{same as those in ggplot functions, typically inherited from ggmap()}

\item{data}{data typically inherited from ggmap().}

\item{geom}{same as those in ggplot2 functions but should be uisng "point".}

\item{position}{same as those in ggplot2 functions.}

\item{na.rm}{same as those in ggplot2 functions.}

\item{show.legend}{same as those in ggplot2 functions.}

\item{inherit.aes}{same as those in ggplot2 functions.}

\item{...}{same as those in ggplot2 functions.}
}
\value{
a ggplot2 object
}
\description{
Plot census data on ggmap
}
