% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_center_and_bbox_from_data.R
\name{get_center}
\alias{get_center}
\title{get coordinate of center from data}
\usage{
get_center(df)
}
\arguments{
\item{df:}{data frame with two columns lon (or long) and lat, usually
generated from a shape file.}
}
\value{
a vector of coordinate c(long, lat)
}
\description{
get coordinate of center from data
}
