% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_center_and_bbox_from_data.R
\name{get_zoom}
\alias{get_zoom}
\title{Get right zoom level for map range of data}
\usage{
get_zoom(df)
}
\arguments{
\item{df:}{data frame with two columns lon (or long) and lat, usually
generated from a shape file.}
}
\value{
an integer
}
\description{
Get right zoom level for map range of data
}
