% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{princeton}
\alias{princeton}
\title{Princeton Color Palette}
\format{
A list of \link[palette:palette]{palette::palette} vectors
}
\usage{
princeton
}
\description{
Princeton Color Palette
}
\examples{
princeton$core
}
\keyword{datasets}
