% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_color_princeton_c}
\alias{scale_color_princeton_c}
\alias{scale_fill_princeton_c}
\alias{scale_color_princeton_b}
\alias{scale_fill_princeton_b}
\alias{scale_color_princeton_d}
\alias{scale_fill_princeton_d}
\alias{scale_colour_princeton_d}
\alias{scale_colour_princeton_c}
\alias{scale_colour_princeton_b}
\title{Scales consistent with Princeton Brand Template}
\usage{
scale_color_princeton_c(palette = "sequential_orange", ...)

scale_fill_princeton_c(
  palette = "sequential_orange",
  na.color = "#5b575d",
  ...
)

scale_color_princeton_b(
  palette = "sequential_orange",
  na.color = "#5b575d",
  ...
)

scale_fill_princeton_b(
  palette = "sequential_orange",
  na.color = "#5b575d",
  ...
)

scale_color_princeton_d(palette = "core", ...)

scale_fill_princeton_d(palette = "core", ...)

scale_colour_princeton_d(palette = "core", ...)

scale_colour_princeton_c(palette = "sequential_orange", ...)

scale_colour_princeton_b(
  palette = "sequential_orange",
  na.color = "#5b575d",
  ...
)
}
\arguments{
\item{palette}{Name of the palette to use. See \link{princeton} for available palettes.}

\item{...}{Additional arguments to pass to \link[ggplot2:binned_scale]{ggplot2::binned_scale} for \verb{_b},
\link[ggplot2:continuous_scale]{ggplot2::continuous_scale} for \verb{_c}, or \link[ggplot2:discrete_scale]{ggplot2::discrete_scale} for \verb{_d}}

\item{na.color}{Color to use for \code{NA} values (only for \verb{_b} scales)}
}
\value{
a \code{ggplot} scale object
}
\description{
Scales consistent with Princeton Brand Template
}
\examples{
library(ggplot2)

ggplot(mpg, aes(cty)) +
  geom_density(aes(fill = factor(cyl)), alpha = 0.8) +
  labs(
    title = 'Density plot',
    subtitle = 'City Mileage Grouped by Number of cylinders',
    caption = 'Source: mpg',
    x = 'City Mileage',
    fill = '# Cylinders'
  ) +
  facet_wrap(~ (hwy > 29)) +
  scale_fill_princeton_d()

}
\concept{princeton}
