% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_boundary.R
\name{stat_boundary}
\alias{stat_boundary}
\title{Draw census boudaries on ggmap}
\usage{
stat_boundary(...)
}
\arguments{
\item{...}{see all parameters in geom_boundary}
}
\description{
The same as geom_boundary()
}
