% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_census.R
\name{stat_census}
\alias{stat_census}
\title{Draw census boudaries on ggmap}
\usage{
stat_census(...)
}
\arguments{
\item{...}{see all parameters in geom_boundary}
}
\description{
The same as geom_boundary()
}
