% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{state_county_zipstart}
\alias{state_county_zipstart}
\title{First two digits of zip code in each county}
\format{A data.table with 172 rows and 2 variables
\describe{
  \item{abbr}{abbreviation of state}
  \item{county}{county name}
  \item{ZCTA5}{first two digits of ZCTA5}
}}
\usage{
state_county_zipstart
}
\description{
First two digits of zip code in each county
}
\keyword{datasets}
