% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_state_fips.R
\docType{data}
\name{state_fips}
\alias{state_fips}
\title{fips of each state}
\format{A data.table with 53 rows and 2 variables
\describe{
  \item{state}{abbreviation of state}
  \item{fips}{fips code of the state}
}}
\usage{
state_fips
}
\description{
fips of each state
}
\keyword{datasets}
