% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{state_zipstart}
\alias{state_zipstart}
\title{First two digits of zip code in each state}
\format{A data.table with 172 rows and 2 variables
\describe{
  \item{abbr}{abbreviation of state}
  \item{ZCTA5}{first two digits of ZCTA5}
}}
\usage{
state_zipstart
}
\description{
First two digits of zip code in each state
}
\keyword{datasets}
