% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_princeton}
\alias{theme_princeton}
\title{A theme based on Princeton University design guidelines}
\usage{
theme_princeton(
  princeton_font = TRUE,
  title_family = NULL,
  base_family = NULL,
  ...
)
}
\arguments{
\item{princeton_font}{Should Princeton fonts be used if available?}

\item{title_family}{Font family for titles. It will search locally for fonts}

\item{base_family}{Font family for base text}

\item{...}{Additional arguments to pass to \code{\link[ggplot2:theme]{ggplot2::theme()}}}
}
\value{
a \code{ggplot2} theme object
}
\description{
A theme based on Princeton University design guidelines
}
\examples{
theme_princeton()
}
