% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_shapefiles.R
\name{download_shapefile}
\alias{download_shapefile}
\title{Download shape file with tigris and convert it to data.table with option
to save as a csv file}
\usage{
download_shapefile(state = NULL, geography, year = 2016, ...)
}
\description{
Download shape file with tigris and convert it to data.table with option
to save as a csv file
}
