% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_shapefiles.R
\name{download_zipcode}
\alias{download_zipcode}
\title{Download zip code
It is very different from others}
\usage{
download_zipcode(starts_with = starts_with, year = year, ...)
}
\description{
Download zip code
It is very different from others
}
