% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_boundary.R
\name{geom_boundary}
\alias{geom_boundary}
\title{Add census boudaries to ggmap}
\usage{
geom_boundary(mapping = NULL, data = NULL, geom = "polygon",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, geometry, data_fill = NULL, year = 2016, ...)
}
\description{
Add census boudaries to ggmap
}
