% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmap_modifiy.R
\name{ggmap}
\alias{ggmap}
\title{Modified ggmap::ggmap() function}
\usage{
ggmap(ggmap, extent = "normal", base_layer, maprange = FALSE,
  legend = "right", padding = 0.02, darken = c(0, "black"), ...)
}
\arguments{
\item{ggmap}{an object of class ggmap (from function get_map)}

\item{extent}{how much of the plot should the map take up?
"normal" (default), "device", or "panel"}

\item{base_layer}{a ggplot(aes(...), ...) call. The same as in the original ggmap().}

\item{maprange}{logical for use with base_layer; should the map
define the x and y limits? The same as in the original ggmap().}

\item{legend}{"left", "right" (default), "bottom", "top",
"bottomleft", "bottomright", "topleft", "topright", "none"
(used with extent = "device"). The same as in the original ggmap().}

\item{padding}{distance from legend to corner of the plot (used
with legend, formerly b). The same as in the original ggmap().}

\item{darken}{vector of the form c(number, color), where number
is in [0, 1] and color is a character string indicating the
color of the darken.  0 indicates no darkening, 1 indicates a
black-out. The same as in the original ggmap().}

\item{...}{... The same as in the original ggmap().}
}
\value{
a ggplot object
}
\description{
The extent is default to "normal" and the xlim and ylim are set in `coord_map()` so
they will not delete data outside the limits in added ggplot layers. All others are the
same as ggmap::ggmap().
}
