% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_path_to_tiger.R
\name{set_path_to_tiger}
\alias{set_path_to_tiger}
\title{Set file path to directory storing downloaded and coverted TIGER data}
\usage{
set_path_to_tiger(path = NULL)
}
\arguments{
\item{path}{path to directory holding all downloaded and converted TIGER data, such as
"E:/my_tiger_data" and "~/my_tiger_data".}
}
\description{
Set file path to directory storing downloaded and coverted TIGER data
}
